Cast User id to integer. If User is not logged in when performing the action, will store user_id as 0.

+++ Model/Behavior/LogActionBehavior.php
@@ -201,7 +201,7 @@ class LogActionBehavior extends ModelBehavior {
 		}
 		$authSession = $this->settings[$Model->alias]['authSession'];
 		$userSession = $this->settings[$Model->alias]['userModel'];
-		return Set::extract($_SESSION, $authSession . '.' . $userSession . '.' . 'id');
+		return (integer) Set::extract($_SESSION, $authSession . '.' . $userSession . '.' . 'id');
 	}
 
 }
